Created etbot package and move_base launch file.
The 'etbot' package is the main package of our robot.
It has the necessary config files (they end with .yaml) for the navigation stack.
The package also contains the launch file 'etbot_config'. This file launches the odometry
node, the rosserial_arduino node, freenect_stack and the static tf broadcaster.

The move_base launch file start the map_server and AMCL.