Expanded funcionality
Regarding the CommandCenter class:
- added several new instance variables
- `publish_reset()` now keeps sending the reset message until
the odom topic values have been reset. Previously this was not the case.
- added a `createTwistMsg(x, z)` method to generate a Twist message
with the specified linear and rotational velocity
- added a `translate()` and `rotate()`method to translate or rotate  the robot a
desired distance or angle. The desired distance or angle is specified by the instance variables `desired_distance` and `desired_angle`
Also of note, the distance and angle calculation is done by using tf2 listeners.
-

Added a `bcolors` class for printing colored text.

Added some new instructions and wrapped some in a try except block.